+++ history.md
@@ -8,6 +8,7 @@
  - 清除`paramsAndCallback`方法
  - 添加新方法`getContext`获取内容信息和`getTitle`获取标题信息
  - 优化`resetScroll`方法
+ - 添加`getConfirmColor`和`setConfirmColor`方法
 
  (v2.1.7) <Date:2019/09/01>
  - `webpack`构建工具替换成`gulp`构建工具